pyproject.toml
[tool.poetry]
name = "online-auctions-backend"
version = "0.1.0"
description = "API for online auction service."
authors = ["Nikita Solouyou <solovev.nikita.05@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
python-dotenv = "^1.0.0"
uvicorn = "^0.23.2"
fastapi-users = { extras = ["sqlalchemy"], version = "^12.1.2" }
asyncpg = "^0.28.0"
alembic = "^1.12.0"
pydantic-settings = "^2.0.3"
fastapi-filter = "^1.0.0"
fastapi-pagination = "^0.12.10"
celery = "^5.3.4"
eventlet = "^0.33.3"
fastapi-mail = "^1.4.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.291"
black = "^23.9.1"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 119
exclude = [
    "migrations",
]

[tool.mypy]
exclude = [
    "migrations",
]
plugins = [
    "pydantic.mypy",
]
